requirements.txt
Pillow==10.2.0
exif==1.6.0
pytest==8.0.0 